% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMS2.R
\docType{package}
\name{KMS2}
\alias{KMS2}
\alias{KMS2-package}
\title{An R client to Amazon Key Management Service}
\description{
This is a simple wrapper around the most important features of the related c++ SDK.
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/kms/AWSKMSClient.html}
}
